Rename `Optimizer` to `Solver`
